requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
Django==4.0.3
django-mathfilters==1.0.0
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1

numpy==1.22.3
Pillow==9.0.1
PyPDF2==2.10.0
python-bidi==0.4.2
pytz==2022.1
reportlab==3.6.11
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
webencodings==0.5.1
xhtml2pdf==0.2.5


please download from given file in base directory

mysqlclient @ file:///D:/python/django/demo_project/mysqlclient-1.4.6-cp39-cp39-win32.whl
